Modularise solver fun, move helpers to header
Move Rcpp C++ 11 plugin to header

Format helper functions

WIP move helper funs to new header
