+++ src/finalsize.cpp
@@ -48,12 +48,11 @@ Eigen::VectorXd final_size_cpp(const double &r0,
   }
 
   // scale demography vector
-  Eigen::VectorXd pp0 = demography_vector / (demography_vector.sum());
+  Eigen::VectorXd pp0 = normalise_demography(demography_vector);
 
   // largest real eigenvalue of the contact matrix
-  Eigen::EigenSolver<Eigen::MatrixXd> es(contact_matrix, false);
-  Eigen::MatrixXd eig_vals = es.eigenvalues().real();
-  double eig_val_max = eig_vals.maxCoeff();
+  double eig_val_max = get_max_real_eigenvalue(contact_matrix);
+  
   // scale the next generation matrix for max eigenvalue = r0
   Eigen::MatrixXd mm0 = r0 * (contact_matrix / eig_val_max);
